refactor: streamline AIChat and AISidebar components by removing unused variables and optimizing imports
